requirements.txt
openai
pydantic
chromadb
python-dotenv
sentence-transformers